# Serialize object fix